MGMT-1099 Host: Added Resetting state
The new state indicates that the agent on the host should restart itself,
then the host should be registered again.
In this commit I also added the new state to the RegisterHost transition
so resetting host will be able to register itself.
Unit tests for the resetting state are included.